cambios rene

+++ src/main/java/BackEndC3/ClinicaOdontologica/entity/Domicilio.java
@@ -1,7 +1,9 @@
+
 package BackEndC3.ClinicaOdontologica.entity;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.Getter;
+import lombok.Setter;
 
 @Getter
 @Setter
@@ -10,7 +12,7 @@ import lombok.*;
 public class Domicilio {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     @Column
     private String calle;
     @Column
@@ -27,7 +29,9 @@ public class Domicilio {
         this.provincia = provincia;
     }
 
-    public Domicilio(Integer id, String calle, Integer numero, String localidad, String provincia) {
+
+
+    public Domicilio(Long id, String calle, Integer numero, String localidad, String provincia) {
         this.id = id;
         this.calle = calle;
         this.numero = numero;
@@ -35,7 +39,8 @@ public class Domicilio {
         this.provincia = provincia;
     }
 
-    public Domicilio(){}
+    public Domicilio() {
+    }
 
     @Override
     public String toString() {
